requirements.txt
flet
quotesgeneratorapi-wrapper
python-dotenv